OEPP-7 Fix clearing of cookies
When running acceptance tests locally cookies should be cleared. It is not possible to do this with via selenium server due to
browser permissions (can not delete cookies set in parent levels).
So to clear cookies _cc file is copied to shop during selenium bootstrap and used when running acceptance tests.