Fix acceptance tests retry counter
Before $_iRetryTimesLeft was used for the whole tests file, not only for one test, which does not make much sense.
Now it is fixed and retry times are per test, not per tests file.
Also variables, which should not be used outside the class was made private.
Retry times count was reduced because of introduced fix.