OETL-2 Add modOxUtilsDate to registry
modOxUtilsDate was registered as a module and setModuleInformation() removed it from active modules. Therefore UNITSetTime method was not available and tests crashed.
It is safer and less hackish solution to add modOxUtilsDate to registry instead.
Also reverted compile directory to old one, as it is still not possible to completely overwrite value specified in config.inc.php.