Check if oxUnitTestCase and oxConfig is loaded
Allow overwriting of sShopDir only when oxUnitTestCase and oxConfig classes are loaded in getShopBasePath.
oxUnitTestCase is included at the very bottom of tests bootstrapping, so shop should be installed and prepared for testing at this point.
However if test is not extending UnitTestCase, but PhpUnit_TestCase instead, oxConfig can be not set yet and cause exceptions on these tests.