ESDEV-3057 Move DatabaseRestorer classes to Services
Added additional DatabaseRestoreToFile class, which is used for selenium tests and implements same interface as others.
Renamed and restructured DatabaseHandler class. It no longer does dumping and restoring of database, but can do importing and exporting of database data.
FileHandler class was added for tasks related to files and directories creation.