Fix AllTestsRunner
Add check if test file was already added. Windows has a bug with loading test files twice, both with uppercased directory and lowercased.
Here all test files is stored in testFiles array and later on checked if test file is already registered with PHPUnit.

This causes a small drawback, as it will not be possible to have tests with the same path, but different letter case in linux.