Fix to show correct failure message when element is present but it should not be "assertElementNotPresent"
Before this fix assertElementNotPresent would fail with default message from method _waitForDisappear.

There is no need to recheck if element is available as method _waitForDisappear does exactly that and fail if not succeed.
That's why it's better to pass failure message to method _waitForDisappear to get message which was formed in original test.