Fix assertion in if statement
PHP first of all does && operation which automatically converts array from $this->getTablesForCleanup() to an true or false.
If $this->getTestConfig()->shouldRestoreAfterUnitTests() returns true if statement passes but next code breaks as it expects array and not boolean true.

Change order of calls, as there is no need to get tables for cleanup if configured to not restore DB.

Related #11